chore: rename file to AppUsageStatsContext
