+++ src/frontend/contexts/AppUsageStatsContext.tsx
